first alot of bug

+++ gamesStoreProduct/lib/controllers/admin.js
@@ -27,14 +27,10 @@ exports.add_property = [
         const data = await db.collection(req.body.collection).find().toArray();
         data.map(async (item) => {
             item[req.body.fieldName] = req.body.defaultValue;
-            await db
-                .collection(req.body.collection)
-                .updateOne({ _id: ObjectId(item._id) }, { $set: item });
+            await db.collection(req.body.collection).updateOne({ _id: ObjectId(item._id) }, { $set: item });
         });
 
-        logger.info(
-            `Added ${req.body.fieldName} to ${req.body.collection} collection.`
-        );
+        logger.info(`Added ${req.body.fieldName} to ${req.body.collection} collection.`);
         res.json({ data, success: true });
     },
 ];
@@ -49,14 +45,10 @@ exports.delete_property = [
         const data = await db.collection(req.body.collection).find().toArray();
         data.map(async (item) => {
             delete item[req.body.fieldName];
-            await db
-                .collection(req.body.collection)
-                .updateOne({ _id: ObjectId(item._id) }, { $set: item });
+            await db.collection(req.body.collection).updateOne({ _id: ObjectId(item._id) }, { $set: item });
         });
 
-        logger.info(
-            `deleted ${req.body.fieldName} to ${req.body.collection} collection.`
-        );
+        logger.info(`deleted ${req.body.fieldName} to ${req.body.collection} collection.`);
         res.json({ data, success: true });
     },
 ];
@@ -68,13 +60,11 @@ exports.clean_token = [
         const db = req.app.get("db");
 
         const tokens = await db.tokens.find().toArray();
-        if (!tokens)
-            return res.status(400).json({ success: false, msg: "No Token!" });
+        if (!tokens) return res.status(400).json({ success: false, msg: "No Token!" });
 
         tokens.map(async (item) => {
             const isExpired = moment().diff(item.expired, "days");
-            if (isExpired > 3)
-                await db.tokens.deleteOne({ _id: ObjectId(item._id) });
+            if (isExpired > 3) await db.tokens.deleteOne({ _id: ObjectId(item._id) });
         });
 
         logger.info("Cleaned all useless tokens.");
@@ -89,10 +79,7 @@ exports.toggle_admin = [
         const db = req.app.get("db");
 
         const { error } = isObjectId("UserId", req.params.id);
-        if (error)
-            return res
-                .status(400)
-                .json({ success: false, msg: error.details[0].message });
+        if (error) return res.status(400).json({ success: false, msg: error.details[0].message });
 
         const user = await db.users.findOne({ _id: ObjectId(req.params.id) });
         if (!user)
@@ -103,9 +90,7 @@ exports.toggle_admin = [
         user.isAdmin = !user.isAdmin;
 
         await db.users.updateOne({ _id: ObjectId(user._id) }, { $set: user });
-        logger.info(
-            `${user.username} has been toggoled admin role: ${user.isAdmin}`
-        );
+        logger.info(`${user.username} has been toggoled admin role: ${user.isAdmin}`);
         res.json({
             success: true,
             msg: `${user.username} have been toggoled admin role.`,
@@ -120,10 +105,7 @@ exports.toggle_developer = [
         const db = req.app.get("db");
 
         const { error } = isObjectId("UserId", req.params.id);
-        if (error)
-            return res
-                .status(400)
-                .json({ success: false, msg: error.details[0].message });
+        if (error) return res.status(400).json({ success: false, msg: error.details[0].message });
 
         const user = await db.users.findOne({ _id: ObjectId(req.params.id) });
         if (!user)
@@ -134,9 +116,7 @@ exports.toggle_developer = [
         user.isDeveloper = !user.isDeveloper;
 
         await db.users.updateOne({ _id: ObjectId(user._id) }, { $set: user });
-        logger.info(
-            `${user.username} has been toggoled developer role: ${user.isDeveloper}`
-        );
+        logger.info(`${user.username} has been toggoled developer role: ${user.isDeveloper}`);
         res.json({
             success: true,
             msg: `${user.username} have been toggoled developer role.`,
